+++ src/Dependencies.lua
@@ -93,6 +93,7 @@ require('src.states.game.PauseMenu')
 require('src.states.game.PopupWindow')
 require('src.states.game.SettingsState')
 require('src.states.game.ShopMenu')
+require('src.states.game.ClosingState')
 
 require('src.states.entity.BaseEntity')
 require('src.states.entity.CoffeeMachine')